create a clients module
