+++ dbt/adapters/bigquery/credentials.py
@@ -4,14 +4,9 @@ from dataclasses import dataclass, field
 from functools import lru_cache
 from typing import Any, Dict, Iterable, Optional, Tuple, Union
 
-from google.api_core.client_info import ClientInfo
-from google.api_core.client_options import ClientOptions
 from google.auth import default
 from google.auth.exceptions import DefaultCredentialsError
 from google.auth.impersonated_credentials import Credentials as ImpersonatedCredentials
-from google.cloud.bigquery.client import Client as BigQueryClient
-from google.cloud.dataproc_v1 import JobControllerClient, BatchControllerClient
-from google.cloud.storage.client import Client as StorageClient
 from google.oauth2.credentials import Credentials as GoogleCredentials
 from google.oauth2.service_account import Credentials as ServiceAccountCredentials
 from mashumaro import pass_through
@@ -22,9 +17,6 @@ from dbt_common.exceptions import DbtConfigError, DbtRuntimeError
 from dbt.adapters.contracts.connection import Credentials
 from dbt.adapters.events.logging import AdapterLogger
 from dbt.adapters.exceptions.connection import FailedToConnectError
-
-import dbt.adapters.bigquery.__version__ as dbt_version
-
 
 _logger = AdapterLogger("BigQuery")
 
@@ -163,53 +155,7 @@ class BigQueryCredentials(Credentials):
         return d
 
 
-def bigquery_client(credentials: BigQueryCredentials) -> BigQueryClient:
-    try:
-        return _bigquery_client(credentials)
-    except DefaultCredentialsError:
-        _logger.info("Please log into GCP to continue")
-        _setup_default_credentials()
-        return _bigquery_client(credentials)
-
-
-def storage_client(credentials: BigQueryCredentials) -> StorageClient:
-    return StorageClient(
-        project=credentials.execution_project,
-        credentials=_get_credentials(credentials),
-    )
-
-
-def job_controller_client(credentials: BigQueryCredentials) -> JobControllerClient:
-    options = ClientOptions(
-        api_endpoint=f"{credentials.dataproc_region}-dataproc.googleapis.com:443",
-    )
-    return JobControllerClient(
-        credentials=_get_credentials(credentials),
-        client_options=options,
-    )
-
-
-def batch_controller_client(credentials: BigQueryCredentials) -> BatchControllerClient:
-    options = ClientOptions(
-        api_endpoint=f"{credentials.dataproc_region}-dataproc.googleapis.com:443",
-    )
-    return BatchControllerClient(
-        credentials=_get_credentials(credentials),
-        client_options=options,
-    )
-
-
-def _bigquery_client(credentials: BigQueryCredentials) -> BigQueryClient:
-    return BigQueryClient(
-        credentials.execution_project,
-        _get_credentials(credentials),
-        location=getattr(credentials, "location", None),
-        client_info=ClientInfo(user_agent=f"dbt-bigquery-{dbt_version.version}"),
-        client_options=ClientOptions(quota_project_id=credentials.quota_project),
-    )
-
-
-def _setup_default_credentials() -> None:
+def setup_default_credentials() -> None:
     try:
         run_cmd(".", ["gcloud", "--version"])
     except OSError as e:
@@ -225,7 +171,7 @@ def _setup_default_credentials() -> None:
     run_cmd(".", ["gcloud", "auth", "application-default", "login"])
 
 
-def _get_credentials(credentials: BigQueryCredentials) -> GoogleCredentials:
+def google_credentials(credentials: BigQueryCredentials) -> GoogleCredentials:
     if credentials.impersonate_service_account:
         return _impersonated_credentials(credentials)
     return _google_credentials(credentials)
